Solucion de algunos errores

+++ Sobrecarga de Funciones/src/sobrecarga/Ej2.java
@@ -5,12 +5,12 @@ import java.util.Scanner;
 public class Ej2 {
 
 	public static double media(int n1, int n2) {
-		double media =(n1 + n2)/2;
+		double media = (double) (n1 + n2)/2;
 		return  media;
 	}
 	
 	public static double media(int n1, int n2, int n3) {
-		double media =(n1 + n2+ n3)/3;
+		double media = (double) (n1 + n2+ n3)/3;
 		return media;
 	}
